start of styling
safe to work on xdai chain first. we have to set a seperate init hash if we want to link in sushiswap amm for mainnet. if we cant we have to make default page farming page. otherwise it will crash if it tries to load swap page on mainnet. have to remove networks we dont use to prevent crashes. WIP